refactor request init + shutdown into request.rs
