+++ otel/tests/phpt/disable/dotenv-disable/dotenv-disable.phpt
@@ -13,5 +13,5 @@ OTEL_TRACES_EXPORTER=memory
 ?>
 --EXPECTF--
 %A
-[%s] [DEBUG] [pid=%d] [ThreadId(%d)] otel: event src/lib.rs:%d message=OpenTelemetry::RINIT: OTEL_SDK_DISABLED is set to true, skipping initialization
+[%s] [DEBUG] [pid=%d] [ThreadId(%d)] %s message=OpenTelemetry::RINIT: OTEL_SDK_DISABLED is set to true, skipping initialization
 %A
