+++ otel/tests/run-server.php
@@ -15,7 +15,7 @@ function run_server(string $file, array $options, string $path = ''): void {
     $cmd = sprintf(
         '%s php %s -S %s:%d -t %s %s > %s 2>&1 & echo $!',
         'OTEL_TRACES_EXPORTER=console OTEL_SPAN_PROCESSOR=simple',
-        '-d extension=otel.so', //-d otel.log.level=debug
+        '-d extension=otel.so -d otel.log.level=warn',
         $host,
         $port,
         escapeshellarg($docRoot),
